feat: deck of cards & table draw (#2)
### Why?

To play the game we need a deck of cards. Ranked by values and suit. To
start off the game we also need players to draw a random card for the
dealer button.

### What's changed?
- Added cards
- Added deck of cards
- Tables can be created out of four players. Where the player who draws
the highest card starts at the north position (with their team m8 across
the table at south). The player who draw the highest card in the
opposite team starts at the east position.